% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{as_quanteda_corpus}
\alias{as_quanteda_corpus}
\title{Convert an NTE tibble to a quanteda corpus}
\usage{
as_quanteda_corpus(data)
}
\arguments{
\item{data}{Corpus tibble with a \code{text} column.}
}
\value{
A \code{quanteda::corpus} object.
}
\description{
Expects a \code{text} column on the input tibble.
}
